Refactor: Split useTransactions hook
Split the useTransactions hook into smaller files for better organization and maintainability.
